+++ test/Feature/CleanRegex/Match/groupBy/MatchPatternTest.php
@@ -282,7 +282,7 @@ class MatchPatternTest extends TestCase
     private function filtered(): GroupByPattern
     {
         return $this->match()
-            ->ignoring(function (Detail $detail) {
+            ->remaining(function (Detail $detail) {
                 $detail->setUserData("verify me:$detail");
                 return !in_array($detail->text(), ['14cm', '13mm']);
             })